requirements.txt
# Core Dependencies
openai>=1.12.0
anthropic>=0.18.0
google-genai>=1.0.0
python-dotenv>=1.0.0

# Video Processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
Pillow>=10.2.0
opencv-python>=4.9.0
numpy>=1.26.0

# Audio
audioxa>=0.1.0
pydub>=0.25.1
edge-tts>=6.1.10

# Web Scraping & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.116.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pytrends>=4.9.2

# AI & ML
transformers>=4.37.0
torch>=2.1.0
sentence-transformers>=2.3.1
clip-interrogator>=0.6.0

# Data Processing
pandas>=2.2.0
pytube>=15.0.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
pyyaml>=6.0.1

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0

# Database & Auth
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.26.0
aiosmtplib>=3.0.0
email-validator>=2.1.0

# Caption/Transcription (Deepgram)
deepgram-sdk>=3.0.0